requirements.txt
django==3.2.20
djangorestframework==3.14.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.8.6
celery==4.4.1
redis==3.4.1
flake8==5.0.4
